requirements.txt
# AI and API clients
openai==1.54.0

# Image processing
pillow==10.4.0
requests==2.32.0

# Web scraping for brand analysis
beautifulsoup4==4.12.3
lxml==5.3.0

# Frontend
streamlit==1.39.0

# Environment variables
python-dotenv==1.0.1

# Data handling
pandas==2.2.0

# JSON handling
jsonschema==4.23.0